+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - Issue #1103: fixed error on tool wfsSearch with `active`=`true`.
 - Issue #1124: fixed url used in WMS-time layer and in zoomTo environment to not use more than one questionmark.
 - Issue #1126: prevent that the "legend"-parameter from config is overwritten by capabilities request
+- Issue #1131: ShadowTool fixed inital language can now also be `de-DE`.
 - Issue #1143: FileImport: open draw tool button is styled correctly.
 - Issue #1149: fixed mobile error for type custom tree portals caused by tree model.
 - Issue #1152: Layertree with background maps in folder structure: When changing a background map, the previously selected background map is deselected.